requirements.txt
Authlib==1.2.1
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.7
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pycparser==2.21
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.0
